requirements.txt
python-dotenv~=1.0.0
pymilvus~=2.3.5
flask~=3.0.0
flask-restx~=1.3.0
nats-py~=2.6.0
sentence-transformers~=2.2.2